# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DFDR Swagger OpenAPI Documentation",
        "description": "Endpoint from DFDR Application",
        "contact": {
            "email": "admin@dfdr.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/language": {
            "get": {
                "tags": [
                    "Language"
                ],
                "description": "View All",
                "operationId": "findAllLanguage",
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Language"
                ],
                "description": "Add new",
                "operationId": "postLanguage",
                "requestBody": {
                    "description": "Add new Language",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "language_name"
                                ],
                                "properties": {
                                    "language_name": {
                                        "description": "Language Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/language/{id}": {
            "get": {
                "tags": [
                    "Language"
                ],
                "description": "View By Id",
                "operationId": "showLanguage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Language",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            },
            "put": {
                "tags": [
                    "Language"
                ],
                "description": "Update Data",
                "operationId": "putLanguage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Language",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Add new Language",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "language_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Update Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Language"
                ],
                "description": "Delete Data",
                "operationId": "destroyLanguage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Language",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Delete Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/language/search/{name}": {
            "get": {
                "tags": [
                    "Language"
                ],
                "description": "Search Data",
                "operationId": "searchLanguage",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Language",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Search Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer -token from login-)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}